requirements.txt
python-dotenv~=1.0.1
numpy==1.23.5
torch==2.3.0
aiohttp~=3.10.3
fastapi~=0.112.1

bcrypt==4.0.1
passlib[bcrypt]~=1.7.4

uvicorn[standard]
SQLAlchemy~=2.0.32
requests~=2.32.3
pyjwt~=2.9.0
paho-mqtt==2.1.0
psycopg2==2.9.9
ultralytics~=8.2.79
ultralytics-thop==0.2.8
openvino
pytesseract
Pillow~=10.4.0
easyocr~=1.7.1
opencv-contrib-python==4.10.0.82
opencv-python==4.7.0.72
norfair~=2.2.0
pyserial~=3.5
python-Levenshtein

cython~=3.0.10
zlib~=1.2.13
future~=1.0.0
pygments~=2.18.0
yaml~=0.2.5
pyyaml~=6.0.2
basicsr~=1.4.2
realesrgan~=0.3.0
tqdm~=4.66.5
setuptools~=72.1.0
torchvision~=0.19.0
gfpgan~=1.3.8
levenshtein~=0.25.1
scipy~=1.14.0
matplotlib~=3.9.2
pydantic~=2.8.2